+++ pyudev/_libudev.py
@@ -86,6 +86,7 @@ SIGNATURES = {
         ref=([udev_p], udev_p),
         get_sys_path=([udev_p], c_char_p),
         get_dev_path=([udev_p], c_char_p),
+        get_run_path=([udev_p], c_char_p),
         get_log_priority=([udev_p], c_int),
         set_log_priority=([udev_p, c_int], None)),
     # enumeration