+++ tests/test_core.py
@@ -46,6 +46,12 @@ def test_context_device_path(context):
     assert context.device_path == '/dev'
 
 
+@pytest.need_udev_version('>= 167')
+def test_context_runpath(context):
+    assert pytest.is_unicode_string(context.run_path)
+    assert context.run_path == '/run'
+
+
 def test_context_log_priority_get(context):
     assert isinstance(context.log_priority, int)
     assert syslog.LOG_EMERG <= context.log_priority <= syslog.LOG_DEBUG